#220 Move players management into core operations
