--- src/MultiplayerMod/Multiplayer/Configuration/MultiplayerConnectionException.cs
@@ -1,7 +1,0 @@
-﻿using MultiplayerMod.Exceptions;
-
-namespace MultiplayerMod.Multiplayer.Configuration;
-
-public class MultiplayerConnectionException : MultiplayerException {
-    public MultiplayerConnectionException(string message) : base(message) { }
-}
