+++ src/MultiplayerMod/Multiplayer/CoreOperations/PlayersManagement/ClientInitializationRequestEvent.cs
@@ -1,7 +1,8 @@
 ﻿using MultiplayerMod.Core.Events;
+using MultiplayerMod.Multiplayer.Players;
 using MultiplayerMod.Network;
 
-namespace MultiplayerMod.Multiplayer.Players.Events;
+namespace MultiplayerMod.Multiplayer.CoreOperations.PlayersManagement;
 
 public record ClientInitializationRequestEvent(
     IMultiplayerClientId ClientId,
