+++ src/MultiplayerMod/Multiplayer/CoreOperations/PlayersManagement/Commands/RequestPlayerStateChangeCommand.cs
@@ -1,8 +1,9 @@
 ﻿using System;
 using MultiplayerMod.Multiplayer.Commands;
+using MultiplayerMod.Multiplayer.Players;
 using MultiplayerMod.Network;
 
-namespace MultiplayerMod.Multiplayer.Players.Commands;
+namespace MultiplayerMod.Multiplayer.CoreOperations.PlayersManagement.Commands;
 
 [Serializable]
 [MultiplayerCommand(Type = MultiplayerCommandType.System)]
