+++ src/MultiplayerMod/Multiplayer/CoreOperations/PlayersManagement/PlayersManagementException.cs
@@ -1,0 +1,7 @@
+﻿using MultiplayerMod.Exceptions;
+
+namespace MultiplayerMod.Multiplayer.CoreOperations;
+
+public class PlayersManagementException : MultiplayerException {
+    public PlayersManagementException(string message) : base(message) { }
+}
