+++ src/MultiplayerMod/Multiplayer/Players/PlayerIdentity.cs
@@ -5,4 +5,6 @@ namespace MultiplayerMod.Multiplayer.Players;
 [Serializable]
 public record PlayerIdentity {
     public Guid Value { get; } = Guid.NewGuid();
+
+    public override string ToString() => Value.ToString();
 }
